==> notebook_1.ipynb <==
import pandas as pd

#### dataset_1 
- Complete Data
- __6027__ Rows

#### dataset_2
- Applied filters to dataset_1
- __4145__ Rows

#### dataset_3 
- Made corrections in dataset_2

df = pd.read_csv("dataset_3.csv")

df.columns

## Column - __Addresses__

#### Data collected about authors : 
- firstname
- lastname
- fullname
- department
- organization
- country

Addresses = df['Addresses']

Addresses[0]

### *getAuthors( )*

def getAuthors(text):
    data = []
    addresses = text.split("[")[1:]
    for i in addresses:
        address = i.replace("\n","").strip();
        if address[-1] == ";":
            address = address[:-1]   
        address = address.split("]")    
        address_parts = address[1].split(",")
        organization = address_parts[0]
        department = address_parts[1]
        country = address_parts[-1].split(" ")[-1]
        authors = address[0].replace(".","").split(";")
        for j in authors:
            author = {}
            name = j.replace("-"," ").split(",",1)
            author["first_name"] = name[-1].strip()
            author["last_name"] = name[0].strip()
            author["full_name"] = author["first_name"] + ", " + author["last_name"]
            author["department"] = department.replace("dept","").replace("Dept","").strip()
            author["organization"] = organization.strip()
            author["country"] = country.strip()
            data.append(author)
    return data

getAuthors(Addresses[0])

## Column - __Leader Addresses__

#### Data collected about leaders : 
- leader name
- leader country

leaderAddresses = df['Leader Addresses']

leaderAddresses[0]

### *getLeader( )*

def getLeader(text):
    data = {}
    data["leader_last_name"] = text.split(",",1)[0]
    data["leader_country"] = text.replace(".","").strip().split(" ")[-1]
    return data

getLeader(leaderAddresses[0])

## Collecting all __authors__ and __leaders__

authors = []
for i in Addresses:
    authors.append(getAuthors(i))
len(authors)

leaders = []
count = 0
for i in leaderAddresses:
    leaders.append(getLeader(i))
len(leaders)

### *addLeaders( )*

def addLeaders(teams_data, leaders_data):
    result = []
    for team,leader in zip(teams_data,leaders_data):
        leader_last_name = leader['leader_last_name']
        leader_name = ""
        temp = []
        for i in team:
            if i['last_name'] == leader_last_name:
                leader_name = i['full_name']
        for j in team:
            j['leader'] = leader_name
            temp.append(j)
        result.append(temp)
    return result

teams_1 = addLeaders(authors,leaders)

### *addCitations( )*

citations = df['Times Cited']

def addCitations(teams,citations):
    result = []
    for team,times_cited in zip(teams,citations):
        temp = []
        for j in team:
            j['times_cited'] = times_cited
            temp.append(j)
        result.append(temp)
    return result

teams_2 = addCitations(teams_1,citations)

## __Removing__ teams without Indians

def remove(teams_data):
    data = []
    for i in teams_data:
        countries = []
        for author in i:
            countries.append(author['country'])
        if("India" in countries):    
            data.append(i)    
    return data

teams_data = remove(teams_2)

### *addTeamIds()*

def addTeamids(teams_data):
    team_id = 1
    for i in teams_data:
        for author in i:
            author['team_id'] = team_id
        team_id += 1

addTeamids(teams_data)

### *addAuthorPositions()*

def addAuthorPositions(teams_data):
    for i in teams_data:
        for index,author in enumerate(i):
            if(index == 0):
                author['position'] = 'first'
            elif(index == len(i)-1):
                author['position'] = 'last'
            else:
                author['position'] = 'other'
            
            if(author['full_name'] == author['leader']):
                if(author['position'] == 'first'):
                    author['position'] = 'first, leader'
                elif(author['position'] == 'last'):
                    author['position'] = 'last, leader'
                else:
                    author['position'] = 'leader'

teams_data[:2]

addAuthorPositions(teams_data)

data = []
for i in teams_data:
    for author in i:
        data.append(author)        

len(data)

## Adding __api names__ to authors

for i in data:
    first_name = i["first_name"].split(" ")
    if(len(first_name[0]) > 1):
        i['api_name'] = first_name[0]
    else:
        last_name = i["last_name"].split(" ")
        i['api_name'] = last_name[0]   

file_data = pd.DataFrame(data)
# file_data.to_csv("data.csv", index=False)

#### Deleted :

- 508 - Team : 248
- A-602 Venkatesh Pooja - Team : 943
