==> notebook_2.ipynb <==
import pandas as pd
import json
from itertools import groupby

df = pd.read_csv("data.csv")
json_data = json.loads(df.to_json(orient="records"))

teams = []

def key_func(k):
    return k['team_id']

for key, value in groupby(json_data, key = key_func):
    teams.append(list(value))

## __Removing__ teams with only Indians

def check(countries):
    x = set(countries)
    if('India' in x):
        x.remove('India')
    if(len(x) > 0):
        return True
    else:
        return False

def exclude(teams_data):
    data = []
    for i in teams_data:
        countries = []
        for author in i:
            countries.append(author['country'])
        if(check(countries) == True):    
            data.append(i)    
    return data

teams_data = exclude(teams)

len(teams_data)

data = []
for i in teams_data:
    for author in i:
        data.append(author)   

len(data)

file_data = pd.DataFrame(data)
# file_data.to_csv("collaboration.csv", index=False)
