==> notebook_3.ipynb <==
import pandas as pd
import json

df = pd.read_csv('data_indians.csv')
organizations = df.organization.unique()

## Adding __organization types__

def organization_type(name):
    name = str(name)
    name = name.lower()
    organization_type = ""
    
    if("inst" in name):
        organization_type = "Institute"
    elif("coll" in name):
        organization_type = "College"
    elif("hosp" in name or "clin" in name):
        organization_type = "Hospital"
    elif("coll" in name and "hosp" in name):
        organization_type = "College and Hospital"
    elif("univ" in name):
        organization_type = "University"
    elif("acad" in name):
        organization_type = "Academy"
    elif("ctr" in name or "res" in name):
        organization_type = "Research Center"
    elif("sch" in name):
        organization_type = "School"
    else:
        organization_type = "None"
        
    return organization_type

data = []

for i in organizations:
    organization = {}
    organization['name'] = i
    organization['type'] = organization_type(i)
        
    data.append(organization)

len(data)

file_data = pd.DataFrame(data)
# file_data.to_csv("organizations.csv", index=False)
