==> notebook_4.ipynb <==
import pandas as pd
import json

## Adding __organization_type__ to __data_indians.csv__

authors_df = pd.read_csv('data_indians.csv')
organizations_df = pd.read_csv('organizations.csv')

authors_data = json.loads(authors_df.to_json(orient="records"))
organizations_data = json.loads(organizations_df.to_json(orient="records"))

for author in authors_data:
    for i in organizations_data:
        if(author['organization'] == i['name']):
            author['organization_type'] = i['type']
print("Done")

file_data = pd.DataFrame(authors_data)
# file_data.to_csv("data_indians.csv", index=False)
