==> notebook_5.ipynb <==
import pandas as pd
import json
from itertools import groupby

df = pd.read_csv("data_indians.csv")
json_data = json.loads(df.to_json(orient="records"))

all_teams = []

def key_func(k):
    return k['team_id']

for key, value in groupby(json_data, key = key_func):
    all_teams.append(list(value))

organizations = df.organization.unique()

### *getPublicationsAndCitations( )*

def getPublicationsAndCitations(organizations, all_teams):
    data = []
    for i in organizations:
        temp = {}
        temp['name'] = i
        pubs_count = 0
        cits_count = 0
        for team in all_teams:
            orgs = set()
            cits = set()
            for author in team:
                orgs.add(author['organization'])
                cits.add(author['times_cited'])
            if(i in orgs):
                pubs_count = pubs_count + 1  
                cits_count = cits_count + (list(cits)[0])
        temp['publications'] = pubs_count
        temp['citations'] = cits_count
        data.append(temp)
        pubs_count = 0
        cits_count = 0
    
    return data

data = getPublicationsAndCitations(organizations, all_teams)
print("Done")

file_data = pd.DataFrame(data)
# file_data.to_csv("publications.csv", index=False)
