==> notebook_6.ipynb <==
## Ratios __MM__, __FF__ and __MF__

import pandas as pd
import json
from itertools import groupby

df = pd.read_csv("data_indians.csv")
df = df[['team_id','gender']]
df = df.groupby(['team_id'])['gender'].apply(set)

df.head()

m_count = 0
f_count = 0
other = 0

m_set = {'male'}
f_set = {'female'}
other_set = {'female', 'male'}

for i in df:
    if i == f_set:
        f_count = f_count + 1
    if i == m_set:
        m_count = m_count + 1
    if i == other_set:
        other = other + 1

result = {}

result['MM'] = m_count
result['FF'] = f_count
result['MF'] = other
result['Total'] = m_count + f_count + other

print(result)
